requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.9
opencv-python>=4.9
numpy>=1.26
pillow>=10.0
pandas>=2.0
scikit-learn>=1.4
matplotlib>=3.8
pytesseract>=0.3.10
torch>=2.2
torchvision>=0.17
tqdm>=4.66
